Move some general structs from protocol redis reader to protocol redis
