+++ src/config.c
@@ -26,7 +26,6 @@
 #include "data_structures/queue_mpmc/queue_mpmc.h"
 #include "slab_allocator.h"
 #include "support/simple_file_io.h"
-#include "protocol/redis/protocol_redis_reader.h"
 #include "module/module.h"
 #include "network/io/network_io_common.h"
 #include "data_structures/hashtable/mcmp/hashtable.h"
