+++ src/module/redis/command/module_redis_command_get.c
@@ -37,8 +37,8 @@
 #include "storage/channel/storage_channel.h"
 #include "storage/db/storage_db.h"
 #include "module/redis/module_redis.h"
-#include "module/redis/module_redis_command.h"
 #include "module/redis/module_redis_connection.h"
+#include "module/redis/module_redis_command.h"
 #include "network/network.h"
 #include "worker/worker_stats.h"
 #include "worker/worker_context.h"
