+++ src/network/channel/network_channel.c
@@ -20,6 +20,7 @@
 #include "xalloc.h"
 #include "config.h"
 #include "module/module.h"
+#include "protocol/redis/protocol_redis.h"
 #include "protocol/redis/protocol_redis_reader.h"
 #include "network/io/network_io_common.h"
 #include "data_structures/hashtable/mcmp/hashtable.h"
