+++ src/protocol/redis/protocol_redis.h
@@ -25,6 +25,18 @@ enum protocol_redis_types {
 };
 typedef enum protocol_redis_types protocol_redis_types_t;
 
+enum protocol_redis_resp_version {
+    PROTOCOL_REDIS_RESP_VERSION_2,
+    PROTOCOL_REDIS_RESP_VERSION_3
+};
+typedef enum protocol_redis_resp_version protocol_redis_resp_version_t;
+
+enum protocol_redis_reader_protocol_types {
+    PROTOCOL_REDIS_READER_PROTOCOL_TYPE_INLINE,
+    PROTOCOL_REDIS_READER_PROTOCOL_TYPE_RESP
+};
+typedef enum protocol_redis_reader_protocol_types protocol_redis_reader_protocol_types_t;
+
 #ifdef __cplusplus
 }
 #endif
