+++ src/protocol/redis/protocol_redis_reader.h
@@ -30,18 +30,6 @@ enum protocol_redis_reader_states {
 };
 typedef enum protocol_redis_reader_states protocol_redis_reader_states_t;
 
-enum protocol_redis_resp_version {
-    PROTOCOL_REDIS_RESP_VERSION_2,
-    PROTOCOL_REDIS_RESP_VERSION_3
-};
-typedef enum protocol_redis_resp_version protocol_redis_resp_version_t;
-
-enum protocol_redis_reader_protocol_types {
-    PROTOCOL_REDIS_READER_PROTOCOL_TYPE_INLINE,
-    PROTOCOL_REDIS_READER_PROTOCOL_TYPE_RESP
-};
-typedef enum protocol_redis_reader_protocol_types protocol_redis_reader_protocol_types_t;
-
 enum protocol_redis_reader_op_type {
     PROTOCOL_REDIS_READER_OP_TYPE_COMMAND_BEGIN,
     PROTOCOL_REDIS_READER_OP_TYPE_COMMAND_END,
